Refactor Simulator into Interface

+++ Assets/Scripts/Components/FireParticleCollision.cs
@@ -2,14 +2,14 @@ namespace OFogo
 {
     public struct FireParticleCollision
     {
-        public int indexA;
-        public int indexB;
+        public short indexA;
+        public short indexB;
         public float distSq;
 
         public FireParticleCollision(int indexA, int indexB, float distSq)
         {
-            this.indexA = indexA;
-            this.indexB = indexB;
+            this.indexA = (short)indexA;
+            this.indexB = (short)indexB;
             this.distSq = distSq;
         }
     }
